factor notebook output routines into separate translation unit
